refactor:ContactWeaponに名前を変更,WeaponBaseにEquip/Unequip処理を移動
